Different wind shaders for trees, hedges and light weight stuff like grass/flowers/wheat etc.
Input handling is better when alt-tabbing and losing focous on the browser window.

BUG FIX: Ground is now seen again...

CollectiblePreview fixes.

Screens fixes for ScreenAdventureLog and ScreenStatistics + Sector4 (can open stations screens again)

+++ src/components/ui/core/CollectiblePreview.tsx
@@ -140,9 +140,9 @@ const CollectiblePreview: React.FC<CollectiblePreviewProps> = ({ type, isLocked,
 
         const renderer = new THREE.WebGLRenderer({
             antialias: true,
-            alpha: true, // Transparent bakgrund så TacticalCard lyser igenom
+            alpha: true,
             powerPreference: 'high-performance',
-            precision: 'mediump'
+            precision: 'highp'
         });
         renderer.setPixelRatio(Math.min(window.devicePixelRatio, 2));
         renderer.setSize(width, height);
